Back : 라우터, 컨트롤러 세팅 완료

+++ server/models/evaluation.js
@@ -13,7 +13,7 @@ module.exports = (sequelize, DataTypes) => {
   }
   Evaluation.init(
     {
-      grade: DataTypes.INTEGER,
+      grade: DataTypes.FLOAT,
       content: DataTypes.STRING,
     },
     {
